+++ chaosazure/webapp/constants.py
@@ -1,0 +1,1 @@
+RES_TYPE_WEBAPP = 'Microsoft.Web/sites'