+++ chaosazure/machine/constants.py
@@ -1,1 +1,5 @@
 RES_TYPE_VM = "Microsoft.Compute/virtualMachines"
+
+# OS Types
+OS_WINDOWS = "windows"
+OS_LINUX = "linux"