+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased][]
 -   Introduce rollback action for virtual machines
+-   Separate service fabric functionality from the chaostoolkit-azure extension
 
 
 [Unreleased]: https://github.com/chaostoolkit-incubator/chaostoolkit-azure/compare/0.3.1...HEAD