+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 -   Introduce rollback action for virtual machines
 -   Separate service fabric functionality from the chaostoolkit-azure extension
 -   Introduce CPU stress action for VMs
+-   Remove not used azure-mgmt dependencies (#53)
 
 
 [Unreleased]: https://github.com/chaostoolkit-incubator/chaostoolkit-azure/compare/0.3.1...HEAD